backend/obscinity_image/requirements.txt
uvicorn[standard]
Pillow
numpy
python-multipart
aiohttp
tensorflow
pymongo
fastapi
gunicorn


